pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.1",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.0",
    "wikipedia>=1.4.0",
]
